moves passado para o módulo snake
